Fix lava for sphere, raindrops generally
Pegging Lava's general noise to a depth-wise parameter meant that it was
basically scrolling through an image, instead of morping and such.
Switching depth and width factors fixes this.
Rain was being weird in a race-conditiony kinda way: pixels getting
stuck, sometimes in different colors, the whole rig flashing a color,
  etc. Calculating the colors once and reusing them appears to fix this.