Basic MIDI input support
Super-simple skeleton to demonstrate how MIDI input can be put on the
display. MidiBus is just one library for MIDI and Processing; it seems
to work fine. The issue is that it needs a PApplet to talk to, that
being the main Equanemone class, so using it in a plugin is weird.

The Equanemone class implements noteOn/noteOff, calls these on the
plugin class, which then draws to its canvas. This ridiculously simple
demo shows how notes could be drawn based on pitch, but there are of
course infinite possibilities!