Add Img2Opc class for outputting to Open Pixel Control
I wrote this class a while back while working on Open Pixel Control; it
is unfortunately mostly-copied from the OPC repository:
https://github.com/zestyping/openpixelcontrol/blob/master/Processing/libraries/img2opc/src/Img2Opc.java
I just could not be bothered to fix the myriad issues compiling and
using the file as a Processing library. Not sure if it's the fault of
Processing, Java, or other, but I got all sorts of errors that just
weren't worth fixing. Including it as another file in this project works
like a charm.

An OPC server will need to be running to accept the pixel data; an
OpenGL server comes with the OPC repository. Using this, framerates for
a large array are greatly increased (~15fps to ~45fps).
It's a bit convoluted, but now I'm confident that we can use Processing
for the near- and even medium-term as Equanemone grows!