OPC layout script; Img2Opc integration tweaks; etc.
As of this change I'm better situated to develop for the 8x40x4 display
hopefully going to The Plunge. opc_layout.py is a quick script to take
width/height/depth and generate an OPC layout file for it in our general
style. Img2Opc is tweaked to more efficiently (and um, correctly)
send the current plugin's image to the OPC server. I did a bunch of
hacky stuff to make the dimensions work out the way I thought they
should, along with tweaks to the GL server itself. These are all highly
arbitrary, sorry. Figuring out the proper way to do all the geometry is
not what I feel like doing right now. Anyway, it works! Super-fast
rendering of the display; should make further development smooth into
the future.