Add wiggle to EchoVideo
Renders 70% of the height of the video. When a tentacle is jostled, the
column for that tentacle bounced up and down for a little while, using
the extra 30% of the frame for the extra bounce space. The bounce has a
size (0.15 = 30%), a decay time, a decay power, and speed.
I'm not entirely satisfied with how it looks, particularly how there's a
lot of little bouncing at the end. I want big bounces, then a sharp
decline. There's probably a great function for that. Log or something.