Exit at time, fix crash situation
Adds checkTimedExit() to exit() if we're within 5 minutes of a specified time.
This will shut down the display at sunrise.
(The laptop should sleep shortly thereafter.)
And don't totally barf if there are fewer strips than we expect: just iterate
over available strips instead. This can happen if 1 of the 2 pixelpushers
isn't reporting for some reason.

I sincerely hope this doesn't add a bunch of terrible line endings to the file.